Rename module to snowplow-demo-compose
