+++ snowplow-demo-compose/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-//    namespace 'com.snowplowanalytics.snowplowdemonew'
+//    namespace 'com.snowplowanalytics.snowplowdemocompose'
     compileSdkVersion 33
 
     defaultConfig {
-        applicationId 'com.snowplowanalytics.snowplowdemonew'
+        applicationId 'com.snowplowanalytics.snowplowdemocompose'
         minSdkVersion 26
         targetSdkVersion 33
         versionCode 1
@@ -48,18 +48,17 @@ dependencies {
     implementation 'androidx.compose.material:material'
     implementation 'androidx.compose.ui:ui-tooling-preview'
     debugImplementation 'androidx.compose.ui:ui-tooling'
-    
     implementation 'androidx.activity:activity-compose:1.5.1'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.5.1'
     implementation "androidx.navigation:navigation-compose:2.5.3"
-    
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.6.0'
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation project(':snowplow-android-tracker')
     implementation 'androidx.preference:preference:1.1.1'
     implementation 'androidx.browser:browser:1.3.0'
+    
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':snowplow-android-tracker')
     
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
     implementation "com.squareup.retrofit2:converter-gson:2.0.0"
