+++ snowplow-demo-compose/proguard-rules.pro
