+++ snowplow-demo-compose/src/main/java/com/snowplowanalytics/snowplowdemocompose/ui/SchemaDetailViewModel.kt
@@ -1,10 +1,10 @@
-package com.snowplowanalytics.snowplow_demo_new.ui
+package com.snowplowanalytics.snowplowdemocompose.ui
 
 import androidx.compose.runtime.*
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.snowplowanalytics.snowplow_demo_new.data.IgluAPIService
-import com.snowplowanalytics.snowplow_demo_new.data.SchemaUrlParts
+import com.snowplowanalytics.snowplowdemocompose.data.IgluAPIService
+import com.snowplowanalytics.snowplowdemocompose.data.SchemaUrlParts
 import kotlinx.coroutines.launch
 import org.json.JSONObject
