+++ snowplow-demo-compose/src/main/java/com/snowplowanalytics/snowplowdemocompose/ui/SchemaListScreen.kt
@@ -1,11 +1,9 @@
-package com.snowplowanalytics.snowplow_demo_new.ui
+package com.snowplowanalytics.snowplowdemocompose.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowForward
@@ -13,16 +11,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.snowplowanalytics.snowplow.event.ScreenView
-import com.snowplowanalytics.snowplow_demo_new.R
-import com.snowplowanalytics.snowplow_demo_new.data.SchemaUrlParts
-import com.snowplowanalytics.snowplow_demo_new.data.Tracking
-import java.util.*
+import com.snowplowanalytics.snowplowdemocompose.R
+import com.snowplowanalytics.snowplowdemocompose.data.SchemaUrlParts
+import com.snowplowanalytics.snowplowdemocompose.data.Tracking
 
 @Composable
 fun SchemaListScreen(
