+++ snowplow-demo-compose/src/main/res/drawable-v24/ic_launcher_foreground.xml
