+++ snowplow-demo-compose/src/main/res/drawable/ic_launcher_snowplow_web.png
