+++ snowplow-demo-compose/src/main/res/mipmap-hdpi/ic_launcher_round.webp
