+++ snowplow-demo-compose/src/main/res/mipmap-xhdpi/ic_launcher.webp
