+++ snowplow-demo-compose/src/main/res/mipmap-xxhdpi/ic_launcher.webp
