+++ snowplow-demo-compose/src/main/res/values-night/themes.xml
