+++ snowplow-demo-compose/src/main/res/values/colors.xml
