+++ snowplow-demo-compose/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">snowplow-demo-new</string>
+    <string name="app_name">snowplow-demo-compose</string>
     <string name="back_button">Back</string>
     <string name="detail_button">Details</string>
 </resources>
