+++ snowplow-demo-compose/src/main/res/values/themes.xml
